backend(admin): extract the admin into a separate app
