+++ firebase.json
@@ -8,8 +8,8 @@
         "function": "generateDeliveryPDF"
       },
       {
-        "source": "/admin/acceptOrganization{,/**}",
-        "function": "acceptOrganization"
+        "source": "/admin{,/**}",
+        "function": "admin"
       },
       {
         "source": "/ip-upload{,/**}",
